feat(one-line-commands): adds simple-commands support to runem
This is quite a big refactor to add one-liners to .runem.yml, improving
the speed of on-boarding.

This means that the quick route to use should be a .runem.yml file that
looks something like:

```yml
- job:
    command: echo "hello world!"
```

instead of needing the following two files
```yml
- job:
    addr:
      file: path/to/file.py
      function: _function_name
    label: Job Label
    when:
      phase: some_phase
      tags:
        - tag
```

```py
from runem.run_command import run_command
from typing import Any

def _function_name(*kwargs:Any) -> None:
  run_command(["echo", "hello world"], **kwargs)
```

So, that's 2 lines instead of 15 so 87% less code just to get started.

Note that the key difference the 'address' and 'command' entries.

Also note that almost all of the options are now optional. This means
that the above function-addressed config can just be the following:

```yml
- job:
    addr:
      file: path/to/file.py
      function: _function_name
    label: Job Label
```

We use 'shlex' to avoid the word splitting problem when parsing commands
to be run.

There are some rough edges around tag handling, that we should tackle
before releasing.