chore(fix-mypy): fixes spurious test-types
This is a heavier "fix" than just a chore, but not really a fix either.

Basically the test was writing an invalid yml file and the code supported
that bad behaviour. This replaces that buggy test with a correctly typed
one and adds some earlier conformance to the config loader for a
(hopefully) minor functional change.