feat(multi-cwd): clones jobs for each cwd path given in .runem.yml
Allows a job to have multiple cwds. That job will then be run once for
each of those cwds.

This make running individual commands across many project path much
easier and faster as the normal is to run them in serial.

This gives us reporting on each path bit as well as faster completion of
tasks.