fix(help-text-test): fixes the help output, again
Adding more switches to runem cli changes the usage line significantly.
This addresses that by only removing the problematic bit __main__ or -c
and not the rest of the switches.

We do this because the rest of the switches might be on the 'usage:'
line or wrapped on the line below.

The differences come from when running via xdist or single-threaded
pytest and would presumably also manifest in other threaded environs