feat(better-job-function-typing): adds stronger typing for job and hook tasks
We do this by typing the kwargs convenience variable, for which we need
to use `Unpack` (for back compatibility), and `type_extensions` for
cross-python-version compatibility (i.e. something that work for all
targeted version of python).

As a side-effect of this we get the benefit of seeing when and where we
add extra data into the call-stack for `job_execute()` and it emerges
that we only extend the key-word args when we are calling hooks in a
non-threaded way, which makes sense.

We do several things to achieve this:
  1. We have common parameters passed to both hooks and job-tasks
      - These common parameters make the hooks and job-task feel similar
        to develop.
  2. We put all hook-specific kwargs in one place, and mark each as
     optional.
      - this, for now, is mainly because we only have one hook, so this
        will likely change.
  3. We share and combine kwargs in a range of inheritance types, mainly
     to work with(/around?) python-typing, which isn't great in this
     type of situation.