feat(file-controls): adds --always-files switch to force files to be checked
Because of how we apply filters, filters for the files are applied
before they're passed to the relevant jobs.

This means that runem can now always check a subset of files, even if
they're not in the HEAD or staged files, meaning we can verify if some
files that we care about for a rebase or similar are copecetic.