feat(file-controls): adds way to get files changed between branches
This is useful if we want to do quick checks only on code that has
changed since 'main' when pushing, or on a sibling branch or similar.

For example, if we are using `--exec` in a rebase, we can start running
check half-way through the list of commits, instead of every commit, but
have the code check all files that have changes - meaning we can see the
impact of changes throughout the branch, without working on *every*
commit - sometimes this is very useful.

NOTE: this switch and the related git switches should be moved into a
      `pluggy` plugin that needs to be designed.