fix(complete-job-count): fixes and simplifies the in-progress report code
The core problem was that the remaining-job-count didn't match the
number of job-labels that were being shown.

The root-cause was due to some of the job-tasks completing before the
`_update_progress()` thread had started monitoring the jobs.

We fix this by adding a new parameter to the job-executer that tracks
when job completed. It's a very simple fix and reduces the overall
complexity for about the same cost of threading-primitives.