fix(module-job): adds .runem.yml dir to sys.path allowing dynamically import when *installed*
The dynamic imports were working locally, where the runem module was
"local" to the `module` dot-path BUT when `runem` was installed as a
thridparty the dynamic import of the `module` path would fail. This
works around that issue by adding the `.runem.yml` dir-path to the
`sys.path`, which means the `module` config type will always work when
the import path is relative to `.runem.yml`, which matches some of the
rest of the code.

We should, at some point, consider allowing configuring a path that can
define an extra sys.path to add, or something neater.